extract carousel into component; color tweaks
