+++ frontend/src/pages/ExploreData/ExploreDataPage.module.scss
@@ -40,6 +40,10 @@
   margin: 20px auto 0 auto;
 }
 
+.NonClickableMadlibText {
+  color: $alt-dark;
+}
+
 .ReportContainer {
   width: 100%;
 }
