+++ frontend/src/pages/ExploreData/ExploreDataPage.tsx
@@ -1,16 +1,13 @@
-import { Box, Grid } from "@material-ui/core";
+import { Box } from "@material-ui/core";
 import NavigateNextIcon from "@material-ui/icons/NavigateNext";
 import React, { useEffect, useState } from "react";
 import { useCookies } from "react-cookie";
 import { STATUS } from "react-joyride";
 import Carousel from "react-material-ui-carousel";
-import { Fips } from "../../data/utils/Fips";
 import ReportProvider from "../../reports/ReportProvider";
 import {
   getMadLibPhraseText,
-  getMadLibWithUpdatedValue,
   getSelectedConditions,
-  insertOptionalThe,
   MadLib,
   MadLibId,
   MADLIB_LIST,
@@ -31,14 +28,14 @@ import {
   useSearchParams,
 } from "../../utils/urlutils";
 import styles from "./ExploreDataPage.module.scss";
-import OptionsSelector from "./OptionsSelector";
-import { useLocation } from "react-router-dom";
 import { srSpeak } from "../../utils/a11yutils";
 import { urlMap } from "../../utils/externalUrls";
 import { VariableConfig } from "../../data/config/MetricConfig";
 import { INCARCERATION_IDS } from "../../data/variables/IncarcerationProvider";
 import useScrollPosition from "../../utils/hooks/useScrollPosition";
 import { useHeaderScrollMargin } from "../../utils/hooks/useHeaderScrollMargin";
+import { useLocation } from "react-router-dom";
+import CarouselMadLib from "./CarouselMadlib";
 
 const Onboarding = React.lazy(() => import("./Onboarding"));
 
@@ -301,62 +298,6 @@ function ExploreDataPage() {
         </div>
       </div>
     </>
-  );
-}
-
-function CarouselMadLib(props: {
-  madLib: MadLib;
-  setMadLib: (updatedMadLib: MadLib) => void;
-}) {
-  // TODO - this isn't efficient, these should be stored in an ordered way
-  function getOptionsFromPhraseSegement(
-    phraseSegment: PhraseSegment
-  ): Fips[] | string[][] {
-    // check first option to tell if phraseSegment is FIPS or CONDITIONS
-    return isNaN(Object.keys(phraseSegment)[0] as any)
-      ? Object.entries(phraseSegment).sort((a, b) => a[0].localeCompare(b[0]))
-      : Object.keys(phraseSegment)
-          .sort((a: string, b: string) => {
-            if (a.length === b.length) {
-              return a.localeCompare(b);
-            }
-            return b.length > a.length ? -1 : 1;
-          })
-          .map((fipsCode) => new Fips(fipsCode));
-  }
-
-  const location = useLocation();
-
-  return (
-    <Grid container justifyContent="center" alignItems="center">
-      <div className={styles.CarouselItem}>
-        {props.madLib.phrase.map(
-          (phraseSegment: PhraseSegment, index: number) => (
-            <React.Fragment key={index}>
-              {typeof phraseSegment === "string" ? (
-                <span style={{ color: "#666" }}>
-                  {phraseSegment}
-                  {insertOptionalThe(props.madLib.activeSelections, index)}
-                </span>
-              ) : (
-                <OptionsSelector
-                  key={index}
-                  value={props.madLib.activeSelections[index]}
-                  onOptionUpdate={(fipsCode: string) => {
-                    props.setMadLib(
-                      getMadLibWithUpdatedValue(props.madLib, index, fipsCode)
-                    );
-                    location.hash = "";
-                    window.scrollTo({ top: 0, behavior: "smooth" });
-                  }}
-                  options={getOptionsFromPhraseSegement(phraseSegment)}
-                />
-              )}
-            </React.Fragment>
-          )
-        )}
-      </div>
-    </Grid>
   );
 }
