+++ frontend/src/pages/ExploreData/OptionsSelector.module.scss
@@ -32,7 +32,7 @@
 }
 
 .MadLibButton {
-  @include inter400();
+  @include inter500();
   text-underline-offset: 3px !important;
   text-decoration: underline !important;
   padding: 0 0.25rem 5px 0.6rem !important;
